Remove PermissionsPluginProxyCallbacks
The callback interface was initially added to let the helper library call the
proxy manager, so it may allow all loaded plugins to interpose on the location
data being passed to the app. However, this interface is messy because there are
too many indirections. A better approach will be implemented in a future commit.

This commit is part of the proxy API design, in
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/7.

Change-Id: Id7c85621d24528d6aa2512a772f5fa578fdb777b