Let the helper library call plugins directly to modify location data
As promised in a459529e89611c248f23b2c7680e249e0105d174, this commit adds a
better way of modifying location data. When the framework loads a plugin that
happens to interpose on location data, it will register the plugin with the
helper via the PermissionsBridge. The helper may then invoke the plugin's
location interposer directly within its location hooks.

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/7.

Change-Id: Ie0c5aae59b407356e0006b767229e081e15938a2