Add the mechanism to load and instantiate a plugin's proxy
This commit adds the methods to load a plugin proxy given the plugin's package
name. If the plugin interposes on the location API, it also informs the helper
library of the plugin's existence, so that the helper library may invoke the
plugins directly within its hooks.

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/7.

Change-Id: If3ed958a64d2fda2323d82611568e9f63b1d8a5d