Refactor how locations are interposed
In apps like Google Maps, it is not easy to modify the location data in plugins
before the app receives the data. See the linked gitlab issue for more
information. This commit attempts to fix this with a more general approach of
modifying location data.

It builds on the following observation: an app receives location data as
instances of the Location class from privileged system services via
Binder. Therefore, by modifying the Location data the moment it is deserialized
from Binder (in Location.createFromParcel(...)), we can modify location data in
a transparent way before an app receives it.

At the moment, it appears this new approach significantly reduces
complexity. The helper library is now reduced to almost nothing, and there is no
need for method intercepts.

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/14.

Change-Id: I5a226ce8c0c7018c819aacbe4f7ec3be96c0d27e