Enable blocking transactions with plugins
We need to call Binder.allowBlocking() on the binder interfaces returned from
the plugin, so that we can make blocking transactions on them. This
necessitates us to write the main PluginService class ourselves instead of
relying on AIDL. Therefore, the plugin API has changed as well.

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/27.

Change-Id: Ia445b45975f9368a0584cb4bb546bfbf7cbf8d05