Correctly perturb data in parcels that contain binder objects
Unlike other objects, binder objects are not serialized into an array of bytes
when written to a parcel (via Parcel.writeStrongBinder()). As evidence, consider
that the Parcel.marshall() method does not work when it contains binder
objects. (See android_os_Parcel_marshall() in android_os_Parcel.cpp). Therefore,
we cannot expect to simply copy binder objects from a given source parcel to
another target parcel. This is a problem when a parcel contains both binder and
perturbable objects.

With this commit, we record binder instances written into a parcel. If the
parcel later turns out to have perturbable objects, we can correctly create the
final parcel with perturbed data by writing from scratch the tracked binder
objects.

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/26.

Change-Id: Ieceac77d39ee6e344a96c675ae3f89ba3086579b