Enable contacts perturbation via plugins
This commits adds a new API to the plugin service and all necessary structures
to enable perturbation of data via the ContentProvider design pattern. It should
be straightforward to extend this to other sensitive data types that also use
content providers (such as text messages).

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/25.

Change-Id: Ia8bd2044d5d80742d1b0eb1cdc3ecf45cdbb3a3a