Implement the initial skeleton of the camera API
This commit sets up the plumbing between the framework and the plugin. It
reports to the plugin when a new camera stream is created and allows the plugin
to respond whether or not it wants to interpose on the stream. In this commit,
the response is simply discarded. The actual interposition feature will be
implemented in a future commit.

This is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/19.