Complete the basic framework required to interpose on camera streams
This commit introduces the InterposableSurface object uses it to interpose on
camera streams. This object accepts as input the app's original surface and
creates a new surface of its own and this new surface is given to the camera
output stream. The camera stream then renders directly into this new surface. On
each new frame rendered into it, the InterposableSurface provides the plugin
with an opportunity to modify the frame before delivering the frame into the
app.

The API to interpose on the frame is not yet provided to the plugin. That will
be added in a future commit.

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/19.