Add onFrameAvailable() method to IPluginCameraInterposer
The onFrameAvailable() method is called by the InterposableSurface each time the
camera stream generates a new frame, and it allows the plugin to modify the
frame before it is given to the app. With this commit, the basic framework for
modifying camera frames is complete. The plugin is now able to perform frame
modifications.

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/19.