Change how the "this" reference is kept in the InterposableSurface
It appears that we have to explicitly pass the "this" reference from Java to
JNI, and make use of that in JNI layer to call back into Java. Using the default
"jobject this" parameter does not work and is the result of several stack
overflows. This commit fixes plugin instabilities and crashes observed with the
OpenCamera application.

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/19.