Add the basic skeleton of the ptrace-based storage tracer
In this commit, zygote forks an additional process for the tracer if the app has
the enableStorageTracer flag set. The tracer uses ptrace to attach to the main
app process and any of the threads it may create. In a future commit, the tracer
will interpose on the "open()" syscall and call the plugin appropriately.

This commit is part of
https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/45.