Do not perturb data when it is transferred between the same process.
While investigating the Yelp issue we found that Dalf intercepts and perturbs data
even when the data transfer happens between the same process. This can happen
when an app's service sends data to its UI.

This commit ensures that Dalf do not perturb the data when it is transferred
between the same process. As a side effect, the Yelp issue is resolved for now.

Issue: https://gitlab.cs.duke.edu/Heimdall/ContextualPermissions/issues/46
Change-Id: I09ef99d96f336b95a6c697c21343d9f432c5fd0a